+++ src/database/index.ts
@@ -10,4 +10,5 @@ export { AbilityService } from './services/AbilityService';
 export { AccountService } from './services/AccountService';
 export { CharacterService } from './services/CharacterService';
 export { CompanionService } from './services/CompanionService';
+export { MobService } from './services/MobService';
 export { ZoneService } from './services/ZoneService';